requirements.txt
langgraph
langchain
langchain-openai
pinecone
python-dotenv
jinja2